# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/sms/ali/send": {
            "get": {
                "description": "发送短信验证码",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "短信验证码"
                ],
                "summary": "发送短信验证码",
                "parameters": [
                    {
                        "type": "string",
                        "description": "手机号",
                        "name": "phone",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/api/sms/smsbao/send": {
            "get": {
                "description": "短信宝发送短信验证码",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "短信验证码"
                ],
                "summary": "短信宝发送短信验证码",
                "parameters": [
                    {
                        "type": "string",
                        "description": "手机号",
                        "name": "phone",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/api/sms/test/send": {
            "get": {
                "description": "发送测试短信验证码",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "短信验证码"
                ],
                "summary": "发送测试短信验证码",
                "parameters": [
                    {
                        "type": "string",
                        "description": "手机号",
                        "name": "phone",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
...
    },
...
}